Share the checksum, git, and record helpers
